pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
  "."
]
[tool.poetry]
name = "ssdlc-proto-02"
version = "0.1.0"
description = ""
authors = ["apodvoisky <apodvoisky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
dependency-injector = "^4.41.0"
flake8 = "^6.0.0"
passlib = "^1.7.4"
envparse = "^0.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pytest-async-sqlalchemy = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
